HaXPath

Documentation

at :: Text -> Expression Text

Access the value of a node's attribute in text form (equivalent to XPath's @).

data Bool

XPath boolean data type.

child :: Node -> RelativePath

The XPath child:: axis.

contains :: Expression Text -> Expression Text -> Expression Bool

The XPath contains() function.

count :: IsPath p => p -> Expression Number

The XPath count() function.

descendantOrSelf :: Node -> RelativePath

The XPath descendant-or-self:: axis.

doubleSlash :: Node -> Path

The XPath // operator.

data Expression t

Opaque representation of an XPath expression.

class Eq t

Type class of XPath types that can be compared for equality.

class Filterable t where

Methods

(#) :: t -> Expression b -> t infixl 3

Filter a set of nodes by the given predicate.

fromRoot :: RelativePath -> Path

Fix a relative path to begin from the document root (i.e. create an absolute path).

class IsPath t where

Type class for allowing XPath-like operations. Do not create instances of this class.

Methods

(./.) :: t -> RelativePath -> t infixl 2

The XPath (non-abbreviated) / operator.

toPath :: t -> Path

Convert to a Path.

class Lit h x | h -> x where

Methods

lit :: h -> Expression x

Create an XPath literal value from a Haskell value.

namedNode :: Text -> Node

Create a node within the given name.

node :: Node

The XPath node() function.

data Node

Opaque representation of an XPath node.

data NodeSet

XPath type representing an unordered set of nodes.

not :: Expression Bool -> Expression Bool

The XPath not(.) function.

data Number

XPath numeric data type.

class Eq t => Ord t

Type class of XPath types that can be ordered.

type Path = Expression NodeSet

The type of XPaths.

position :: Expression Number

The XPath position() function.

data RelativePath

A relative XPath, i.e. an XPath that is relative to the current node.

show :: IsPath p => p -> Text

Display an XPath expression. This is useful to sending the XPath expression to a separate XPath evaluator e.g. a web browser.

text :: Expression Text

The XPath text() function.

data Text

XPath textual (string) data type.

(&&.) :: Expression Bool -> Expression Bool -> Expression Bool infixr 3

The XPath and operator.

(/.) :: IsPath p => p -> Node -> p infixl 2

The XPath abbreviated / operator.

(//.) :: IsPath p => p -> Node -> p infixl 2

The XPath // operator.

(/=.) :: Eq a => Expression a -> Expression a -> Expression Bool infix 4

The XPath != operator.

(<.) :: Ord a => Expression a -> Expression a -> Expression Bool infix 4

The XPath < operator.

(<=.) :: Ord a => Expression a -> Expression a -> Expression Bool infix 4

The XPath <= operator.

(=.) :: Eq a => Expression a -> Expression a -> Expression Bool infix 4

The XPath = operator.

(>.) :: Ord a => Expression a -> Expression a -> Expression Bool infix 4

The XPath > operator.

(>=.) :: Ord a => Expression a -> Expression a -> Expression Bool infix 4

The XPath >= operator.

(||.) :: Expression Bool -> Expression Bool -> Expression Bool infixr 2

The XPath or operator.