HaXPath.Schematic

Documentation

abs :: Expression Number a -> Expression Number a

Schematic XPath equivalent of abs

type Attribute a = Expression Text '[a]

type Bool = Expression Bool '[]

child :: MultiNode s n -> RelativePath s n

contains :: Union a b c => Expression Text a -> Expression Text b -> Expression Bool c

count :: IsPath p u => p s n -> Expression Number '[]

descendantOrSelf :: MultiNode s n -> RelativePath s n

doubleSlash :: MultiNode s n -> Path s n

data Expression (x :: *) (a :: [*])

An XPath expression returning a value of type x, involving zero or more attributes a.

class Filterable (t :: * -> [*] -> *) where

Methods

(#) :: NodesAttributes n a => t s n -> Expression Bool a -> t s n infixl 3

fromInteger :: Integer -> Expression Number '[]

Schematic XPath equivalent of fromInteger

fromRoot :: RelativePath s n -> Path s n

class IsPath u => IsPath (t :: * -> [*] -> *) (u :: *) | t -> u where

Type class for allowing XPath-like operations. Do not create instances of this class.

Methods

toNonSchematicPath :: t s n -> u

Convert a schematic XPath to its non-schematic equivalent.

unsafeFromNonSchematicPath :: u -> t s n

Unsafely (without type checking) convert a non-schematic XPath to its schematic equivalent.

lit :: Lit h x => h -> Expression x '[]

namedNode :: Text -> Node s n

negate :: Expression Number a -> Expression Number a

Schematic XPath equivalent of negate

type Node (s :: *) (n :: *) = MultiNode s '[n]

class NodeAttribute n a

Witnesses that a node of type n may have an attribute of type a.

class NodeAttributes (n :: *) (a :: [*])

Witnesses that a node of type n may have zero or more of a set of attributes a.

class NodesAttributes (n :: [*]) (a :: [*])

Witnesses that a set of nodes n may have zero or more of a set of attributes a.

node :: SchemaNodes s n => MultiNode s n

The XPath node() function.

not :: Expression Bool a -> Expression Bool a

The XPath not() function.

type Number = Expression Number '[]

data Path (s :: *) (n :: [*])

The type of XPaths for a schema s returning a set of nodes which may any of the type-list n.

position :: Expression Number '[]

data RelativePath (s :: *) (n :: [*])

A relative XPath for a schema s returning a set of nodes which may any of the type-list n.

class SchemaNodes (s :: *) (n :: [*]) | s -> n

A given schema s may have use a universe of nodes n.

show :: IsPath p u => p s n -> Text

signum :: Expression Number a -> Expression Number a

Schematic XPath equivalent of 'P.signum

text :: Expression Text '[]

The XPath text() function.

type Text = Expression Text '[]

class Union (l :: [*]) (m :: [*]) (lm :: [*]) | l m -> lm

The union of two sets of types.

unsafeAt :: Text -> Attribute a

Access the value of a node's attribute in text form (equivalent to XPath's @). Unsafe because it has no way to check if the attribute name provided matches the schema. It is recommend to call this function once only for each attribute in your schema. The resulting value can be re-used.

(+.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath + operator.

(-.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath - operator.

(*.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath * operator.

(&&.) :: Union a b c => Expression Bool a -> Expression Bool b -> Expression Bool c infixr 3

The XPath and operator.

(./.) :: IsPath p u => p s m -> RelativePath s n -> p s n infixl 2

(/.) :: IsPath p u => p s m -> MultiNode s n -> p s n infixl 2

(//.) :: IsPath p u => p s m -> MultiNode s n -> p s n infixl 2

(/=.) :: (Eq x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath != operator.

(<.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath < operator.

(<=.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath <= operator.

(=.) :: (Eq x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath = operator.

(>.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath > operator.

(>=.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath >= operator.

(||.) :: Union a b c => Expression Bool a -> Expression Bool b -> Expression Bool c infixr 2

The XPath or() function.