HaXPath.Schematic

Documentation

(&&.) :: Union a b c => Expression Bool a -> Expression Bool b -> Expression Bool c infixr 3

The XPath and operator.

(*.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath * operator.

(+.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath + operator.

(-.) :: Union a b c => Expression Number a -> Expression Number b -> Expression Number c infixl 6

The XPath - operator.

(./.) :: IsPath p u => p s m -> RelativePath s n -> p s n infixl 2

The XPath (non-abbreviated) / operator.

(/.) :: IsPath p u => p s m -> MultiNode s n -> p s n infixl 2

The XPath abbreviated / operator.

(//.) :: IsPath p u => p s m -> MultiNode s n -> p s n infixl 2

The XPath // operator.

(/=.) :: (Eq x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath != operator.

(<.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath < operator.

(<=.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath <= operator.

(=.) :: (Eq x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath = operator.

(>.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath > operator.

(>=.) :: (Ord x, Union a b c) => Expression x a -> Expression x b -> Expression Bool c infix 4

The XPath >= operator.

(||.) :: Union a b c => Expression Bool a -> Expression Bool b -> Expression Bool c infixr 2

The XPath or operator.

abs :: Expression Number a -> Expression Number a

Schematic XPath equivalent of abs

ancestor :: MultiNode s n -> RelativePath s n

The XPath ancestor:: axis.

at :: Text -> Attribute a

Access the value of a node's attribute in text form (equivalent to XPath's @). It is recommend to call this function once only for each attribute in your schema. The resulting value can be re-used.

type Attribute a = Expression Text '[a]

The type of an attribute.

type Bool = Expression Bool '[]

The type of simple boolean expressions.

child :: MultiNode s n -> RelativePath s n

The XPath child:: axis.

contains :: Union a b c => Expression Text a -> Expression Text b -> Expression Bool c

The XPath contains() function.

count :: IsPath p u => p s n -> Number

The XPath count() function.

descendant :: MultiNode s n -> RelativePath s n

The XPath descendant:: axis.

descendantOrSelf :: MultiNode s n -> RelativePath s n

The XPath descendant-or-self:: axis.

doubleSlash :: MultiNode s n -> Path s n

The XPath // operator.

data Expression (x :: *) (a :: [*])

An XPath expression returning a value of type x, involving zero or more attributes a.

class Filterable (t :: * -> [*] -> *) where

Methods

(#) :: NodesAttributes n a => t s n -> Expression Bool a -> t s n infixl 3

Filter a set of nodes by the given predicate.

following :: MultiNode s n -> RelativePath s n

The XPath following:: axis.

followingSibling :: MultiNode s n -> RelativePath s n

The XPath following-sibling:: axis.

fromInteger :: Integer -> Number

Schematic XPath equivalent of fromInteger

fromRoot :: RelativePath s n -> Path s n

Fix a relative path to begin from the document root (i.e. create an absolute path).

class IsPath u => IsPath (t :: * -> [*] -> *) (u :: *) | t -> u where

Type class for allowing XPath-like operations. Do not create instances of this class.

Methods

toNonSchematicPath :: t s n -> u

Convert a schematic XPath to its non-schematic equivalent.

unsafeFromNonSchematicPath :: u -> t s n

Unsafely (without type checking) convert a non-schematic XPath to its schematic equivalent.

lit :: Lit h x => h -> Expression x '[]

Create a literal XPath value.

data MultiNode (s :: *) (n :: [*])

The type of an XPath expression for returning nodes of a variety of types.

namedNode :: Text -> Node s n

Create a node with the given name.

negate :: Expression Number a -> Expression Number a

Schematic XPath equivalent of negate

type Node (s :: *) (n :: *) = MultiNode s '[n]

Type of a single XPath node.

node :: SchemaNodes s n => MultiNode s n

The XPath node() function.

class NodeAttribute n a

Witnesses that a node of type n may have an attribute of type a.

class NodeAttributes (n :: *) (a :: [*])

Witnesses that a node of type n may have zero or more of a set of attributes a.

class NodesAttributes (n :: [*]) (a :: [*])

Witnesses that a set of nodes n may have zero or more of a set of attributes a.

not :: Expression Bool a -> Expression Bool a

The XPath not() function.

type Number = Expression Number '[]

The type of simple numeric expressions.

parent :: MultiNode s n -> RelativePath s n

The XPath parent:: axis.

data Path (s :: *) (n :: [*])

The type of XPaths for a schema s returning a set of nodes which may be any of the type-list n.

position :: Number

The XPath position() function.

data RelativePath (s :: *) (n :: [*])

A relative XPath for a schema s returning a set of nodes which may be any of the type-list n.

class SchemaNodes (s :: *) (n :: [*]) | s -> n

A given schema s may be comprised of a universe of nodes n.

show :: IsPath p u => p s n -> Text

Display an XPath expression. This is useful for sending the XPath expression to a separate XPath evaluator e.g. a web browser.

signum :: Expression Number a -> Expression Number a

Schematic XPath equivalent of signum

text :: Text

The XPath text() function.

type Text = Expression Text '[]

The type of simple text expressions.

class Union (l :: [*]) (m :: [*]) (lm :: [*]) | l m -> lm

The union of two sets of types.